+++ stringmatch/ratio.py
@@ -4,14 +4,24 @@ import Levenshtein
 class Ratio:
     """Contains functions for calculating the ratio of similarity between two strings."""
 
-    def __init__(self):
-        self.scorers = {
+    def __init__(self, scorer: str = "levenshtein") -> None:
+        """Initialize the Ratio class with the correct scoring algorithm.
+
+        Parameters
+        ----------
+        scorer : str, optional
+            _description_, by default "levenshtein"
+        """
+        self.available_scorers = {
             "levenshtein": Levenshtein.ratio,
             "jaro": Levenshtein.jaro,
             "jaro_winkler": Levenshtein.jaro_winkler,
         }
 
-    def ratio(self, string1: str, string2: str, scorer: str = "levenshtein") -> int:
+        # if the scorer is not found, use levenshtein as the default.
+        self.scorer = self.available_scorers.get(scorer, Levenshtein.ratio)
+
+    def ratio(self, string1: str, string2: str) -> int:
         """Returns the similarity score between two strings.
 
         Parameters
@@ -20,12 +30,6 @@ class Ratio:
             The first string to compare.
         string2 : str
             The second string to compare.
-        scorer : str, optional
-            The scorer to use, by default "levenshtein".
-            Available scorers:
-                "levenshtein"
-                "jaro"
-                "jaro_winkler"
 
 
         Returns
@@ -34,18 +38,10 @@ class Ratio:
             The score between 0 and 100.
         """
 
-        if scorer not in self.scorers:
-            scorer = "levenshtein"
-
         # if either string is empty we wanna return 0
-        if not string1 or not string2:
-            return 0
+        return round(self.scorer(string1, string2) * 100) if string1 and string2 else 0
 
-        return round(self.scorers[scorer](string1, string2) * 100)
-
-    def ratio_list(
-        self, string: str, string_list: list[str], scorer: str = "levenshtein"
-    ) -> list[int]:
+    def ratio_list(self, string: str, string_list: list[str]) -> list[int]:
         """Returns the similarity score between a string and a list of strings.
 
         Parameters
@@ -54,16 +50,10 @@ class Ratio:
             The string to compare.
         string_list : list[str]
             The list of strings to compare to.
-        scorer : str, optional
-            The scorer to use, by default "levenshtein".
-            Available scorers:
-                "levenshtein"
-                "jaro"
-                "jaro_winkler"
 
         Returns
         -------
         list[int]
             The scores between 0 and 100.
         """
-        return [self.ratio(string, s, scorer=scorer) for s in string_list]
+        return [self.ratio(string, s) for s in string_list]